Remove non-production code
