+++ MatterControlApplication.cs
@@ -378,38 +378,6 @@ namespace MatterHackers.MatterControl
 			return instance;
 		}
 
-		public static void WriteTestGCodeFile()
-		{
-			using (StreamWriter file = new StreamWriter("PerformanceTest.gcode"))
-			{
-				//int loops = 150000;
-				int loops = 150;
-				int steps = 200;
-				double radius = 50;
-				Vector2 center = new Vector2(150, 100);
-
-				file.WriteLine("G28 ; home all axes");
-				file.WriteLine("G90 ; use absolute coordinates");
-				file.WriteLine("G21 ; set units to millimeters");
-				file.WriteLine("G92 E0");
-				file.WriteLine("G1 F7800");
-				file.WriteLine("G1 Z" + (5).ToString());
-				WriteMove(file, center);
-
-				for (int loop = 0; loop < loops; loop++)
-				{
-					for (int step = 0; step < steps; step++)
-					{
-						Vector2 nextPosition = new Vector2(radius, 0);
-						nextPosition.Rotate(MathHelper.Tau / steps * step);
-						WriteMove(file, center + nextPosition);
-					}
-				}
-
-				file.WriteLine("M84     ; disable motors");
-			}
-		}
-
 		public void LaunchBrowser(string targetUri)
 		{
 			UiThread.RunOnIdle(() =>
@@ -517,7 +485,8 @@ namespace MatterHackers.MatterControl
 		}
 
 		bool closeHasBeenConfirmed = false;
-		bool closeMessageBoxIsOpen = false;
+		internal bool closeMessageBoxIsOpen = false;
+
 		private void ConditionalyCloseNow(bool continueWithShutdown)
 		{
 			// Response received, record that we are not waiting anymore.
@@ -607,44 +576,6 @@ namespace MatterHackers.MatterControl
 			base.OnLoad(args);
 		}
 
-		private static void HtmlWindowTest()
-		{
-			try
-			{
-				SystemWindow htmlTestWindow = new SystemWindow(640, 480);
-				string htmlContent = "";
-				if (true)
-				{
-					string releaseNotesFile = Path.Combine("C:\\Users\\lbrubaker\\Downloads", "test1.html");
-					htmlContent = File.ReadAllText(releaseNotesFile);
-				}
-				else
-				{
-					WebClient webClient = new WebClient();
-					htmlContent = webClient.DownloadString("http://www.matterhackers.com/s/store?q=pla");
-				}
-
-				HtmlWidget content = new HtmlWidget(htmlContent, RGBA_Bytes.Black);
-				content.AddChild(new GuiWidget()
-				{
-					HAnchor = HAnchor.Absolute,
-					VAnchor = VAnchor.Stretch
-				});
-				content.VAnchor |= VAnchor.Top;
-				content.BackgroundColor = RGBA_Bytes.White;
-				htmlTestWindow.AddChild(content);
-				htmlTestWindow.BackgroundColor = RGBA_Bytes.Cyan;
-				UiThread.RunOnIdle((state) =>
-				{
-					htmlTestWindow.ShowAsSystemWindow();
-				}, 1);
-			}
-			catch
-			{
-				int stop = 1;
-			}
-		}
-
 		public override void OnMouseMove(MouseEventArgs mouseEvent)
 		{
 			// run this first to make sure a child has the chance to take the drag drop event
@@ -698,11 +629,6 @@ namespace MatterHackers.MatterControl
 					(new System.Media.SoundPlayer(mediaStream)).Play();
 				}
 			}
-		}
-
-		private static void WriteMove(StreamWriter file, Vector2 center)
-		{
-			file.WriteLine("G1 X" + center.x.ToString() + " Y" + center.y.ToString());
 		}
 
 		private void CheckOnPrinter()
@@ -788,4 +714,3 @@ namespace MatterHackers.MatterControl
 		}
 	}
 }
-
