+++ Tests/MatterControl.Tests/MatterControl.Tests.csproj
@@ -77,6 +77,7 @@
     <Compile Include="MatterControl\TranslationsTests.cs" />
     <Compile Include="MatterControl\UIFieldTestWindow.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RemovedFromProduct.cs" />
     <Compile Include="SceneTests.cs" />
   </ItemGroup>
   <ItemGroup>
